add model Property and method GetLessonProperties
