+++ LessonMonitor/LessonMonitor.API/Models/Lesson.cs
@@ -8,7 +8,12 @@ namespace LessonMonitor.API.Models
     public class Lesson
     {
         public int LessonId { get; set; }
-        public DateTime DateAdd { get; set; }
+        /// <summary>
+        /// Дата проведения занятия
+        /// </summary>
+        public DateTime DateStart { get; set; }
+        [Description("Ключевые слова")]
+        public List<String> Tags { get; set; }
         public string Tittle { get; set; }
         public string Url { get; set; }
